KPD cmd

+++ app.js
@@ -169,6 +169,24 @@ client.on('message', async(message) => {
     if (env == 'PROD' && message.channel.id == id.channels['change-logs']) await message.crosspost().catch(console.error);
     if (env !== 'PROD' && message.content.startsWith(`${config.prefix}execute`) && (message.author.id == id.users.jytesh || message.author.id == id.users.jj || message.author.id == id.users.ej) && message.channel.id == id.channels['bunker-bot-commands']) evald(message);
 
+    var cmdToRun = '';
+
+    if (message.content.startsWith(`${config.prefix}staff`)) {
+        let isStaff = false;
+        // eslint-disable-next-line no-return-assign
+        client.roles.staff.forEach(role => { if (message.member.roles.cache.has(role)) return isStaff = true; });
+        if (isStaff) {
+            message.content = message.content.substring(message.content.indexOf(' ') + 1);
+            switch(message.content.split(' ')[0]) {
+                case 'kpd':
+                    cmdToRun = 'kpd';
+                    break;
+            }
+        }
+    }
+
+    if (cmdToRun != '') client.commands.get(`${cmdToRun}`).run(client, message);
+
     client.setTimeout(async() => {
         if (env == 'PROD' && !message.deleted) {
             if (message.type == 'PINS_ADD' && message.author.id == client.user.id) message.delete();
@@ -243,15 +261,18 @@ client.on('message', async(message) => {
                 if (isStaff) {
                     message.content = message.content.substring(message.content.indexOf(' ') + 1);
                     switch(message.content.split(' ')[0]) {
-                        case 'rule':
-                            cmdToRun = 'rules';
-                            break;
                         case 'email':
                         case 'emails':
                             cmdToRun = 'emails';
                             break;
+                        case 'kpd':
+                            cmdToRun = 'kpd';
+                            break;
                         case 'roles':
                             cmdToRun = 'roles';
+                            break;
+                        case 'rule':
+                            cmdToRun = 'rules';
                             break;
                     }
                 }
